requirements.txt
Flask==1.1.2
gensim==3.8.3
requests==2.25.1
google-api-python-client
google-cloud-bigquery
google-cloud-language
grpcio